refactor: directory구조 변경
