+++ src/style/style.css
@@ -155,23 +155,24 @@ button:hover {
 }
 
 .editPage_editor {
-    /*
+    
     caret-color: rgba(255, 255, 255, 0.81);
     width: 100%;
     display: flex;
     flex-direction: column;
     position: relative;
     align-items: center;
-    flex-grow: 1;*/
-    display: flex;
+    flex-grow: 1;
+    /* display: flex;
     flex-direction: column;
+    align-items: center;
     /* background: rgb(166, 181, 81); */
-    padding: 10px;
+    padding: 10px; 
 }
 
 .title,
 .content {
-    width: 100%;
+    width: 708px;
     padding: 10px;
     color: #D4D4D4;
     background-color: inherit;
@@ -206,9 +207,9 @@ button:hover {
     color: white;
     border-radius: 4px;
     font-size: 14px;
-    padding: 5px 10px;
     cursor: pointer;
-    margin: 22px;
+    margin-bottom:20px;
+    margin-left: 10px;
 }
 
 .editPage_footer button:hover {
